+++ index.js
@@ -387,10 +387,16 @@ exports.isOwnEmpty = function(obj) {
  * A non-plain object has a `constructor` property set to anything but `Object`.
  * That's the case when you do, for example, `new MyModel`, `new Date`.
  *
+ * `Array.prototype` is not considered a plain object just like an array isn't
+ * a plain object. JavaScript is a prototypical language and the prototype of
+ * an array should be considered an array.
+ *
  * @example
  * Oolong.isPlainObject({name: "John", age: 42}) // => true
  * Oolong.isPlainObject(Object.create(null)) // => true
  * Oolong.isPlainObject(Math) // => true
+ * Oolong.isPlainObject([]) // => false
+ * Oolong.isPlainObject(Array.prototype) // => false
  * Oolong.isPlainObject(new Date) // => false
  * Oolong.isPlainObject("John") // => false
  *
@@ -401,6 +407,7 @@ exports.isOwnEmpty = function(obj) {
 exports.isPlainObject = function(obj) {
   if (obj == null) return false
   if (typeof obj != "object") return false
+  if (Array.isArray(obj)) return false
 
   var prototype = Object.getPrototypeOf(obj)
   if (prototype === null) return true