+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Unreleased
+- Fixes [`Oolong.isPlainObject`][isPlainObject] to return `false` for
+  `Array.prototype`.  
+  JavaScript is a prototypical language after all and the prototype of an array
+  is an array.
+
 ## 1.11.0 (Jun 14, 2015)
 - Adds [`Oolong.lookupGetter`][lookupGetter].
 - Adds [`Oolong.lookupSetter`][lookupSetter].